Use Super Lube on the camera lenses to make them easier to rotate
This helps a recurring issue noted by users during their build. Using a small amount of the Super Lube included with the kit makes it significantly easier to rotate the camera lenses.

Editor's Note: Since this content is repeated twice, it probably makes sense to eventually pull this into its own step (like we're doing with the tensioners in #15).