adds vacuum sensor interposer board information to docs
adds new section called maintenance and upgrades to hold the rev03 motherboard vacuum sensor interposer installation instructions. also updates the actuator page to mention the interposers, link to the installation guide, and let users with the interposer know that the gain in the GCode command should be adjusted from G4 to G2.